started flock tests

+++ RTSNavigationLib/Source/RTSNavigationLib/Flocking/Flock.cpp
@@ -6,7 +6,7 @@
 #include "Boid.h"
 
 namespace RTSPathingLib {
-  Flock::Flock(const glm::ivec2& area_, size_t maxInfluenceRadius, size_t maxNumberOfBoids_) {
+  Flock::Flock(const glm::ivec2& area_, double maxInfluenceRadius, size_t maxNumberOfBoids_) {
     maxNumberOfBoids = maxNumberOfBoids_;
     area             = area_;
     cellSize         = maxInfluenceRadius;
